+++ app/scripts/lib/lforms-util.js
@@ -319,6 +319,7 @@ LForms.Util = {
    */
   setFHIRContext: function(fhirContext) {
     LForms.fhirContext = fhirContext;
+    LForms.fhirCapabilities = {}; // our own flags about what the server can do
     delete LForms._serverFHIRReleaseID; // in case the version changed
   },
 
@@ -582,10 +583,13 @@ LForms.Util = {
   /**
    * Get a formatted date string from a date object
    * for example: "2016-10-31T14:42:12Z"
-   * @param objDate a date object
+   * @param objDate a date object, or a valid string representation of date.
    * @returns a formatted date string
    */
   dateToDTMString: function(objDate) {
+    if(typeof objDate === 'string') {
+      objDate = LForms.Util.stringToDate(objDate, true);
+    }
     return objDate.toISOString();
   },
 
@@ -600,19 +604,19 @@ LForms.Util = {
     if(! strDate || (typeof strDate) != 'string') { // maybe already a date object.
       return strDate;
     }
-    
+
     if(strDate.trim() === 't') {
       return new Date();
     }
-    
+
     let m = moment(strDate, parseDateFormats, true);
     if(looseParsing && !m.isValid()) { // Make another attempt for loose parsing.
       m = moment(strDate);
     }
     return m.isValid() ? m.toDate() : null;
   },
-  
-  
+
+
   /**
    * Validate date object or date string. If string, check to see if it is in acceptable formats.
    * See stringToDate() for acceptable formats.
@@ -622,8 +626,8 @@ LForms.Util = {
   isValidDate: function(date) {
     return !!(LForms.Util.stringToDate(date));
   },
-  
-  
+
+
   /**
    * Format a date object with given format. Refer to momentjs documentation for
    * format specification.
@@ -635,8 +639,8 @@ LForms.Util = {
   formatDate: function(date, format) {
     return moment(date).format(format);
   },
-  
-  
+
+
   /**
    * Check if an item's value is empty, where the data has no meaningful use.
    * @param value the value to be tested